terraform/ml_pipeline/functions/requirements.txt
functions-framework==3.*
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.42.1
google-cloud-pubsub==2.18.0
google-cloud-logging==3.8.0
requests==2.31.0
pyyaml==6.0